requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.1
dj-database-url==1.0.0
Django==4.0.5
django-crispy-forms==1.14.0
django-heroku==0.3.1
django-restframework==0.0.1
djangorestframework==3.13.1
idna==3.3
psycopg2==2.9.3
pycryptodome==3.15.0
pytz==2022.1
requests==2.28.1
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.12
whitenoise==6.2.0
